requirements.txt
# Core
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-huggingface>=0.0.3
langchain-chroma>=0.1.0
langchain-text-splitters>=0.0.1

# Vector store
chromadb>=0.4.0

# Embeddings
sentence-transformers>=2.2.0

# LLM
groq>=0.4.0

# PDF processing
pypdf>=3.0.0

# UI
streamlit>=1.30.0

# Utilities
python-dotenv>=1.0.0
